requirements.txt
nltk==3.5
numpy==1.19.0
Pillow==7.1.2
scipy==1.4.1
six==1.15.0
tensorflow==1.14.0
tflearn==0.3.2


